# src/Playarr.Api.V3/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Playarr",
    "description": "Playarr API docs - The v3 API docs apply to both v3 and v4 versions of Playarr. Some functionality may only be available in v4 of the Playarr application.",
    "license": {
      "name": "GPL-3.0",
      "url": "https://github.com/Playarr/Playarr/blob/develop/LICENSE"
    },
    "version": "3.0.0"
  },
  "servers": [
    {
      "url": "{protocol}://{hostpath}",
      "variables": {
        "protocol": {
          "default": "http",
          "enum": [
            "http",
            "https"
          ]
        },
        "hostpath": {
          "default": "localhost:9797"
        }
      }
    }
  ],
  "paths": {
...
    "/api/v3/customformat/bulk": {
      "put": {
        "tags": [
          "CustomFormat"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomFormatBulkResource"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomFormatResource"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "CustomFormat"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomFormatBulkResource"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CustomFormatBulkResource": {
        "type": "object",
        "properties": {
          "ids": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "includeCustomFormatWhenRenaming": {
            "type": "boolean",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CustomFormatResource": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "includeCustomFormatWhenRenaming": {
            "type": "boolean",
            "nullable": true
          },
          "specifications": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CustomFormatSpecificationSchema"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CustomFormatSpecificationSchema": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "implementation": {
            "type": "string",
            "nullable": true
          },
          "implementationName": {
            "type": "string",
            "nullable": true
          },
          "infoLink": {
            "type": "string",
            "nullable": true
          },
          "negate": {
            "type": "boolean"
          },
          "required": {
            "type": "boolean"
          },
          "fields": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Field"
            },
            "nullable": true
          },
          "presets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CustomFormatSpecificationSchema"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Field": {
        "type": "object",
        "properties": {
          "order": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "label": {
            "type": "string",
            "nullable": true
          },
          "unit": {
            "type": "string",
            "nullable": true
          },
          "helpText": {
            "type": "string",
            "nullable": true
          },
          "helpTextWarning": {
            "type": "string",
            "nullable": true
          },
          "helpLink": {
            "type": "string",
            "nullable": true
          },
          "value": {
            "nullable": true
          },
          "type": {
            "type": "string",
            "nullable": true
          },
          "advanced": {
            "type": "boolean"
          },
          "selectOptions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SelectOption"
            },
            "nullable": true
          },
          "selectOptionsProviderAction": {
            "type": "string",
            "nullable": true
          },
          "section": {
            "type": "string",
            "nullable": true
          },
          "hidden": {
            "type": "string",
            "nullable": true
          },
          "privacy": {
            "$ref": "#/components/schemas/PrivacyLevel"
          },
          "placeholder": {
            "type": "string",
            "nullable": true
          },
          "isFloat": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "PrivacyLevel": {
        "enum": [
          "normal",
          "password",
          "apiKey",
          "userName"
        ],
        "type": "string"
      },
...
      "SelectOption": {
        "type": "object",
        "properties": {
          "value": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "order": {
            "type": "integer",
            "format": "int32"
          },
          "hint": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "X-Api-Key": {
        "type": "apiKey",
        "description": "Apikey passed as header",
        "name": "X-Api-Key",
        "in": "header"
      },
      "apikey": {
        "type": "apiKey",
        "description": "Apikey passed as query parameter",
        "name": "apikey",
        "in": "query"
      }
    }
  },
  "security": [
    {
      "X-Api-Key": [ ]
    },
    {
      "apikey": [ ]
    }
  ]
}